Move most remaining arith helpers
