+++ pandas/core/arrays/timedelta.py
@@ -10,7 +10,8 @@ from pandas._libs.tslibs.timedeltas import array_to_timedelta64
 
 from pandas import compat
 
-from pandas.core.dtypes.common import _TD_DTYPE, _ensure_int64
+from pandas.core.dtypes.common import (
+    _TD_DTYPE, _ensure_int64, is_timedelta64_dtype)
 from pandas.core.dtypes.generic import ABCSeries
 from pandas.core.dtypes.missing import isna
 
@@ -58,7 +59,11 @@ class TimedeltaArrayMixin(DatetimeLikeArrayMixin):
         if values.dtype == np.object_:
             values = array_to_timedelta64(values)
         if values.dtype != _TD_DTYPE:
-            values = _ensure_int64(values).view(_TD_DTYPE)
+            if is_timedelta64_dtype(values):
+                # non-nano unit
+                values = values.astype(_TD_DTYPE)
+            else:
+                values = _ensure_int64(values).view(_TD_DTYPE)
 
         result = object.__new__(cls)
         result._data = values
@@ -91,6 +96,38 @@ class TimedeltaArrayMixin(DatetimeLikeArrayMixin):
         assert other is not NaT
         raise TypeError("cannot subtract a datelike from a {cls}"
                         .format(cls=type(self).__name__))
+
+    def _add_delta(self, delta):
+        """
+        Add a timedelta-like, Tick, or TimedeltaIndex-like object
+        to self.
+
+        Parameters
+        ----------
+        delta : timedelta, np.timedelta64, Tick, TimedeltaArray, TimedeltaIndex
+
+        Returns
+        -------
+        result : same type as self
+
+        Notes
+        -----
+        The result's name is set outside of _add_delta by the calling
+        method (__add__ or __sub__)
+        """
+        if isinstance(delta, (Tick, timedelta, np.timedelta64)):
+            new_values = self._add_delta_td(delta)
+        elif isinstance(delta, TimedeltaArrayMixin):
+            new_values = self._add_delta_tdi(delta)
+        elif is_timedelta64_dtype(delta):
+            # ndarray[timedelta64] --> wrap in TimedeltaArray/Index
+            delta = type(self)(delta)
+            new_values = self._add_delta_tdi(delta)
+        else:
+            raise TypeError("cannot add the type {0} to a TimedeltaIndex"
+                            .format(type(delta)))
+
+        return type(self)(new_values, freq='infer')
 
     def _evaluate_with_timedelta_like(self, other, op):
         if isinstance(other, ABCSeries):
